refactor: move bearer auth utility to src/utils
